fix(DoctorPage): ✨ doctor with multiple phone nums
Creates new component <PageInfoPhones/>

FIX: #127
